+++ src/main/scala/com/raquo/laminar/receivers/ChildReceiver.scala
@@ -1,13 +1,14 @@
 package com.raquo.laminar.receivers
 
 import com.raquo.airstream.core.Observable
-import com.raquo.laminar.nodes.{ReactiveChildNode, ReactiveElement}
+import com.raquo.laminar.nodes.ReactiveElement
 import com.raquo.laminar.setters.ChildSetter
+import com.raquo.laminar.setters.ChildrenSetter.Child
 import org.scalajs.dom
 
 class ChildReceiver(element: ReactiveElement[dom.Element]) {
 
-  def <--($node: Observable[ReactiveChildNode[dom.Node]]): Unit = {
+  def <--($node: Observable[Child]): Unit = {
     (ChildReceiver <-- $node)(element)
   }
 }
@@ -18,7 +19,7 @@ object ChildReceiver {
 
   val text: TextChildReceiver.type = TextChildReceiver
 
-  def <--($node: Observable[ReactiveChildNode[dom.Node]]): ChildSetter = {
+  def <--($node: Observable[Child]): ChildSetter = {
     new ChildSetter($node)
   }
 }