+++ src/main/scala/com/raquo/laminar/receivers/MaybeChildReceiver.scala
@@ -1,24 +1,21 @@
 package com.raquo.laminar.receivers
 
 import com.raquo.airstream.core.Observable
-import com.raquo.laminar.nodes.{ReactiveChildNode, ReactiveElement}
+import com.raquo.laminar.nodes.ReactiveElement
+import com.raquo.laminar.setters.ChildrenSetter.Child
 import com.raquo.laminar.setters.MaybeChildSetter
 import org.scalajs.dom
 
 class MaybeChildReceiver(element: ReactiveElement[dom.Element]) {
 
-  import MaybeChildReceiver.MaybeChildNode
-
-  def <--($node: Observable[MaybeChildNode]): Unit = {
+  def <--($node: Observable[Option[Child]]): Unit = {
     (MaybeChildReceiver <-- $node)(element)
   }
 }
 
 object MaybeChildReceiver {
 
-  type MaybeChildNode = Option[ReactiveChildNode[dom.Node]]
-
-  def <--($maybeChildNode: Observable[MaybeChildReceiver.MaybeChildNode]): MaybeChildSetter = {
+  def <--($maybeChildNode: Observable[Option[Child]]): MaybeChildSetter = {
     new MaybeChildSetter($maybeChildNode)
   }
 }