+++ src/main/scala/com/raquo/laminar/setters/ChildSetter.scala
@@ -3,20 +3,21 @@ package com.raquo.laminar.setters
 import com.raquo.airstream.core.Observable
 import com.raquo.domtypes.generic.Modifier
 import com.raquo.laminar.DomApi
-import com.raquo.laminar.nodes.{ReactiveChildNode, ReactiveComment, ReactiveElement}
+import com.raquo.laminar.nodes.{ReactiveComment, ReactiveElement}
+import com.raquo.laminar.setters.ChildrenSetter.Child
 import org.scalajs.dom
 
-class ChildSetter(nodeObservable: Observable[ReactiveChildNode[dom.Node]])
+class ChildSetter(nodeObservable: Observable[Child])
   extends Modifier[ReactiveElement[dom.Element]] {
 
   override def apply(parentNode: ReactiveElement[dom.Element]): Unit = {
-    var childNode: ReactiveChildNode[dom.Node] = new ReactiveComment("")
+    var childNode: Child = new ReactiveComment("")
 
     // @TODO[Performance] In case of Signal we're doing append(comment)+replace(node), but we could do just one append(node)
     parentNode.appendChild(childNode)(DomApi.treeApi)
     parentNode.subscribe(nodeObservable)(onNext)
 
-    @inline def onNext(newChildNode: ReactiveChildNode[dom.Node]): Unit = {
+    @inline def onNext(newChildNode: Child): Unit = {
       parentNode.replaceChild(childNode, newChildNode)(DomApi.treeApi)
       childNode = newChildNode
     }