+++ src/test/scala/com/raquo/laminar/ChildrenCommandReceiverSpec.scala
@@ -3,7 +3,6 @@ package com.raquo.laminar
 import com.raquo.domtestutils.matching.{ExpectedNode, Rule}
 import com.raquo.laminar.api.L._
 import com.raquo.laminar.collection.CollectionCommand.{Append, Insert, Prepend, Remove, Replace, ReplaceAll}
-import com.raquo.laminar.setters.ChildrenCommandSetter.ChildrenCommand
 import com.raquo.laminar.utils.UnitSpec
 
 import scala.collection.mutable